requirements.txt
Django==1.8.4
django-debug-toolbar==1.3.2
django-grappelli==2.7.1
django-localflavor==1.1
djangorestframework==3.2.1
funcsigs==0.4
mock==1.3.0
pbr==1.4.0
psycopg2==2.6.1
PyYAML==3.11
selenium==2.47.1
six==1.9.0
sqlparse==0.1.16